api/scrum/requirements.txt
coverage==4.2
Django==1.10
django-appconf==1.0.2
django-compressor==2.1
django-crispy-forms==1.6.0
django-extensions==1.7.5
django-filter==1.0.1
django-markdown-deux==1.0.5
django-nose==1.4.4
django-pagedown==0.1.1
djangorestframework==3.5.3
docopt==0.6.2
jedi==0.9.0
Markdown==2.6.7
markdown2==2.3.1
nose==1.3.7
Pillow==3.1.0
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptpython==0.36
Pygments==2.1.3
rcssmin==1.0.6
requests==2.12.4
rjsmin==1.0.12
six==1.10.0
wcwidth==0.1.7
Werkzeug==0.11.13
